Add FAQ section
This PR adds the FAQ section to the website. There's currently an issue with the FAQ header SVG copied from Figma, as depicted in the screenshot below. I'm not sure at the moment why it's displayed differently from it really looks (try the SVG in an online SVG viewer and it displays correctly).